LayoutTests/fast/frames/seamless/seamless-form-post.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

window.addEventListener('message', function(evt) {
    shouldBeEqualToString("frames[0].document.body.textContent", "DONE\n");

    if (window.layoutTestController)
        layoutTestController.notifyDone();
}, false);
</script>
<iframe srcdoc="
    Waiting...
    <script>
        window.onload=function() {
            frames[0].document.getElementsByTagName('form')[0].submit();
        }
    </script>
    <iframe seamless srcdoc='<form method=POST action=resources/done.html>
                                 <input type=submit>
                             </from>
    '></iframe>
"></iframe>
